simple_requirements.txt
pydantic-ai==0.0.20
tree-sitter-languages==1.10.2
GitPython==3.1.44